# restapi/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger Json",
    "description": "This is a generated swagger json for your sails project",
    "termsOfService": "http://example.com/terms",
    "contact": {
      "name": "Theophilus Omoregbee",
      "url": "http://github.com/theo4u",
      "email": "theo4u@ymail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:1337/"
    }
  ],
...
  "paths": {
...
    "/api/billet/reserver": {
      "post": {
        "summary": "/api/billet/reserver",
        "tags": [
          "Billet"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/billet/liste/{evenement_id}": {
      "get": {
        "summary": "/api/billet/liste/{evenement_id}",
        "tags": [
          "Billet"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "evenement_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Route pattern variable `evenement_id`"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/billet/imprimer/{reservation_id}": {
      "get": {
        "summary": "/api/billet/imprimer/{reservation_id}",
        "tags": [
          "Billet"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Route pattern variable `reservation_id`"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}